moves and adapts uri unit tests from token tests to uri-registry tests
